+++ test/libsolidity/syntaxTests/functionTypes/external_function_type_to_address.sol
@@ -1,5 +1,5 @@
 contract C {
     function f() public view returns (address) {
-        return address(this.f);
+        return this.f.address;
     }
 }