+++ test/libsolidity/syntaxTests/structs/address_member_declaration.sol
@@ -1,0 +1,5 @@
+contract C {
+    struct S { uint address; }
+}
+// ----
+// ParserError: (33-40): Expected identifier but got 'address'